Bug fixes and improvements
- Added OnSuccess and OnFaliure pins to SetEIKPresence
- Fixed issues in FindSessionByID, CreateEIKLobby, CreateEIKSession, FindEIKSessions.
FindSessionByID: Preveusly broadcasted: FOnlineSessionSearchResult, now it uses the eik standard: FSessionFindStruct. Fixed issue when it returned both faliure and success, when it should return faliure.
CreateEIKLobby: Preveusly broadcasted session name, we dont need to do this because if the user makes a variable and sets LobbyId to it, it would first set to the correct lobby id, and shortly after setting it to the session name.
CreateEIKSession: preveusly broadcasted empty string after the session id, this would cause the same issue as CreateEIKLobby had.
FindEIKSessions: Fixed issue where it would continue executing even if bWasSuccessful is false.